+++ Sources/VCore/Helpers/UIVIew NSLayoutConstriant Helpers/NSLayoutConstraint.HorizontalAttribute.swift
@@ -23,30 +23,13 @@ extension NSLayoutConstraint {
         /// Trailing edge of the object’s alignment rectangle.
         case trailing
         
-        /// Safe leading edge of the object’s alignment rectangle.
-        case safeLeading
-        
-        /// Safe center along the x-axis of the object’s alignment rectangle.
-        case safeCenterX
-        
-        /// Safe trailing edge of the object’s alignment rectangle.
-        case safeTrailing
-        
         // MARK: Properties
         /// Converts `HorizontalAttribute` to `Attribute`.
         public var toAttribute: Attribute {
             switch self {
-            case .leading, .safeLeading: return .leading
-            case .centerX, .safeCenterX: return .centerX
-            case .trailing, .safeTrailing: return .trailing
-            }
-        }
-        
-        /// Indicates if attribute is constrained to `safeAreaLayoutGuide`.
-        public var isSafe: Bool {
-            switch self {
-            case .leading, .centerX, .trailing: return false
-            case .safeLeading, .safeCenterX, .safeTrailing: return true
+            case .leading: return .leading
+            case .centerX: return .centerX
+            case .trailing: return .trailing
             }
         }
     }
@@ -63,7 +46,9 @@ extension UIView {
     ///         view1.constraintLeading(to: view2),
     ///
     ///         view3.constraintLeading(
+    ///             on: .safeArea,
     ///             to: view4,
+    ///             layoutGuide: .safeArea,
     ///             attribute: .leading,
     ///             relation: .equal,
     ///             constant: 0,
@@ -73,18 +58,20 @@ extension UIView {
     ///     ])
     ///
     public func constraintLeading(
-        to view: UIView,
+        on selfLayoutGuide: UILayoutGuideType? = nil,
+        to view: UIView?,
+        layoutGuide: UILayoutGuideType? = nil,
         attribute: NSLayoutConstraint.HorizontalAttribute = .leading,
         relation: NSLayoutConstraint.Relation = .equal,
         constant: CGFloat = 0,
         multiplier: CGFloat = 1,
         priority: UILayoutPriority? = nil
     ) -> NSLayoutConstraint {
-        .init(
-            item: self.layoutItem(isSafe: false),
+        return .init(
+            item: selfLayoutGuide?.toLayoutGuide(in: self) ?? self,
             attribute: .leading,
             relatedBy: relation,
-            toItem: view.layoutItem(isSafe: attribute.isSafe),
+            toItem: view.flatMap { layoutGuide?.toLayoutGuide(in: $0) } ?? view,
             attribute: attribute.toAttribute,
             multiplier: multiplier,
             constant: constant,
@@ -101,7 +88,9 @@ extension UIView {
     ///         view1.constraintCenterX(to: view2),
     ///
     ///         view3.constraintCenterX(
+    ///             on: .safeArea,
     ///             to: view4,
+    ///             layoutGuide: .safeArea,
     ///             attribute: .centerX,
     ///             relation: .equal,
     ///             constant: 0,
@@ -111,7 +100,9 @@ extension UIView {
     ///     ])
     ///
     public func constraintCenterX(
-        to view: UIView,
+        on selfLayoutGuide: UILayoutGuideType? = nil,
+        to view: UIView?,
+        layoutGuide: UILayoutGuideType? = nil,
         attribute: NSLayoutConstraint.HorizontalAttribute = .centerX,
         relation: NSLayoutConstraint.Relation = .equal,
         constant: CGFloat = 0,
@@ -119,10 +110,10 @@ extension UIView {
         priority: UILayoutPriority? = nil
     ) -> NSLayoutConstraint {
         .init(
-            item: self.layoutItem(isSafe: false),
+            item: selfLayoutGuide?.toLayoutGuide(in: self) ?? self,
             attribute: .centerX,
             relatedBy: relation,
-            toItem: view.layoutItem(isSafe: attribute.isSafe),
+            toItem: view.flatMap { layoutGuide?.toLayoutGuide(in: $0) } ?? view,
             attribute: attribute.toAttribute,
             multiplier: multiplier,
             constant: constant,
@@ -139,7 +130,9 @@ extension UIView {
     ///         view1.constraintTrailing(to: view2),
     ///
     ///         view3.constraintTrailing(
+    ///             on: .safeArea,
     ///             to: view4,
+    ///             layoutGuide: .safeArea,
     ///             attribute: .trailing,
     ///             relation: .equal,
     ///             constant: 0,
@@ -149,7 +142,9 @@ extension UIView {
     ///     ])
     ///
     public func constraintTrailing(
-        to view: UIView,
+        on selfLayoutGuide: UILayoutGuideType? = nil,
+        to view: UIView?,
+        layoutGuide: UILayoutGuideType? = nil,
         attribute: NSLayoutConstraint.HorizontalAttribute = .trailing,
         relation: NSLayoutConstraint.Relation = .equal,
         constant: CGFloat = 0,
@@ -157,124 +152,10 @@ extension UIView {
         priority: UILayoutPriority? = nil
     ) -> NSLayoutConstraint {
         .init(
-            item: self.layoutItem(isSafe: false),
+            item: selfLayoutGuide?.toLayoutGuide(in: self) ?? self,
             attribute: .trailing,
             relatedBy: relation,
-            toItem: view.layoutItem(isSafe: attribute.isSafe),
-            attribute: attribute.toAttribute,
-            multiplier: multiplier,
-            constant: constant,
-            priority: priority
-        )
-    }
-    
-    /// Constraints `UIView`'s `DimensionAttribute.safeLeading` to another `UIView`'s `HorizontalAttribute`,
-    /// with given `relation`, `constant`, `multiplier`, and `priority`.
-    ///
-    /// By default, another `UIView`'s `HorizontalAttribute` is set to `safeLeading`.
-    ///
-    ///     NSLayoutConstraint.activate([
-    ///         view1.constraintSafeLeading(to: view2),
-    ///
-    ///         view3.constraintSafeLeading(
-    ///             to: view4,
-    ///             attribute: .safeLeading,
-    ///             relation: .equal,
-    ///             constant: 0,
-    ///             multiplier: 1,
-    ///             priority: .defaultHigh
-    ///         )
-    ///     ])
-    ///
-    public func constraintSafeLeading(
-        to view: UIView,
-        attribute: NSLayoutConstraint.HorizontalAttribute = .safeLeading,
-        relation: NSLayoutConstraint.Relation = .equal,
-        constant: CGFloat = 0,
-        multiplier: CGFloat = 1,
-        priority: UILayoutPriority? = nil
-    ) -> NSLayoutConstraint {
-        .init(
-            item: self.layoutItem(isSafe: true),
-            attribute: .leading,
-            relatedBy: relation,
-            toItem: view.layoutItem(isSafe: attribute.isSafe),
-            attribute: attribute.toAttribute,
-            multiplier: multiplier,
-            constant: constant,
-            priority: priority
-        )
-    }
-    
-    /// Constraints `UIView`'s `DimensionAttribute.safeCenterX` to another `UIView`'s `HorizontalAttribute`,
-    /// with given `relation`, `constant`, `multiplier`, and `priority`.
-    ///
-    /// By default, another `UIView`'s `HorizontalAttribute` is set to `safeCenterX`.
-    ///
-    ///     NSLayoutConstraint.activate([
-    ///         view1.constraintSafeCenterX(to: view2),
-    ///
-    ///         view3.constraintSafeCenterX(
-    ///             to: view4,
-    ///             attribute: .safeCenterX,
-    ///             relation: .equal,
-    ///             constant: 0,
-    ///             multiplier: 1,
-    ///             priority: .defaultHigh
-    ///         )
-    ///     ])
-    ///
-    public func constraintSafeCenterX(
-        to view: UIView,
-        attribute: NSLayoutConstraint.HorizontalAttribute = .safeCenterX,
-        relation: NSLayoutConstraint.Relation = .equal,
-        constant: CGFloat = 0,
-        multiplier: CGFloat = 1,
-        priority: UILayoutPriority? = nil
-    ) -> NSLayoutConstraint {
-        .init(
-            item: self.layoutItem(isSafe: true),
-            attribute: .centerX,
-            relatedBy: relation,
-            toItem: view.layoutItem(isSafe: attribute.isSafe),
-            attribute: attribute.toAttribute,
-            multiplier: multiplier,
-            constant: constant,
-            priority: priority
-        )
-    }
-    
-    /// Constraints `UIView`'s `DimensionAttribute.safeTrailing` to another `UIView`'s `HorizontalAttribute`,
-    /// with given `relation`, `constant`, `multiplier`, and `priority`.
-    ///
-    /// By default, another `UIView`'s `HorizontalAttribute` is set to `safeTrailing`.
-    ///
-    ///     NSLayoutConstraint.activate([
-    ///         view1.constraintSafeTrailing(to: view2),
-    ///
-    ///         view3.constraintSafeTrailing(
-    ///             to: view4,
-    ///             attribute: .safeTrailing,
-    ///             relation: .equal,
-    ///             constant: 0,
-    ///             multiplier: 1,
-    ///             priority: .defaultHigh
-    ///         )
-    ///     ])
-    ///
-    public func constraintSafeTrailing(
-        to view: UIView,
-        attribute: NSLayoutConstraint.HorizontalAttribute = .safeTrailing,
-        relation: NSLayoutConstraint.Relation = .equal,
-        constant: CGFloat = 0,
-        multiplier: CGFloat = 1,
-        priority: UILayoutPriority? = nil
-    ) -> NSLayoutConstraint {
-        .init(
-            item: self.layoutItem(isSafe: true),
-            attribute: .trailing,
-            relatedBy: relation,
-            toItem: view.layoutItem(isSafe: attribute.isSafe),
+            toItem: view.flatMap { layoutGuide?.toLayoutGuide(in: $0) } ?? view,
             attribute: attribute.toAttribute,
             multiplier: multiplier,
             constant: constant,