+++ CHANGELOG.md
@@ -19,11 +19,16 @@ Views and ViewControllers
 - `InfiniteScrollingCollectionViewDelegate` is renamed to `InfiniteScrollingUICollectionViewDelegate`
 - `ScrollableView` is renamed to `ScrollableUIView`
 
+Helpers
+
+- `UIView` helper `NSLayoutConstraint` API has been updated
+
 Extensions
 
 - `Double.rounded(fractions:)` and `Double.round(fractions:)` methods are added
 - `String.diacriticInsensitiveString(locale:)` method is added
 - `UIViewController.isRootViewController` and `UIViewController.isNonRootViewController` properties are added
+- `NSLayoutConstraint.init(item:attribute:relatedBy:toItem:attribute:multiplier:constant:priority)` initializer is added
 
 ### [3.14.0(45)](https://github.com/VakhoKontridze/VCore/releases/tag/3.14.0) — *2022 07 12*
 
@@ -36,7 +41,7 @@ Views and ViewControllers
 
 Extensions
 
-- `NSAttributedString.init(attributedStrings:)` is added
+- `NSAttributedString.init(attributedStrings:)` initializer is added
 
 ### [3.13.2(44)](https://github.com/VakhoKontridze/VCore/releases/tag/3.13.2) — *2022 07 12*